// src/types/Synergetic/Attendance/iSynVAttendancesWithAbsence.ts
type iSynVAttendancesWithAbsence = {
  AttendanceSeq: number;
  FileType: string;
  FileYear: number;
  FileSemester: number;
  ClassCampus: string;
  ClassCode: string;
  ClassDescription: string | null;
  ID: number;
  AttendanceDate: Date | string;
  AttendancePeriod: number;
  PossibleAbsenceCode: string;
  PossibleAbsenceDescription: string;
  PossibleDescription: string;
  AttendedFlag: boolean;
  PossibleReasonCode: string;
  PossibleReasonDescription: string | null;
  StaffID: number;
  SchoolStaffCode: string | null;
  StaffNameInternal: string | null;
  StaffNameExternal: string | null;
  StaffLegalFullName: string | null;
  StaffSurname: string | null;
  StaffPreferred: string | null;
  StaffPreferredFormal: string | null;
  StaffName: string | null;
  EventDate: Date | string | null;
  AbsenceDate: Date | string | null;
  AbsencePeriod: string | null;
  AbsencePeriodSort: number | null;
  AbsencePeriodDescription: string | null;
  AbsenceTypeCode: string | null;
  AbsenceType: string | null;
  AbsenceTypeDescription: string | null;
  AbsenceTypeReportCode: string | null;
  AbsenceReasonCode: string | null;
  AbsenceReason: string | null;
  AbsenceReasonDescription: string | null;
  EventComment: string | null;
  AbsenceDescription: string | null;
  EnteredInAdvanceFlag: boolean | null;
  AbsenceAdvanceNotifiedFlag: boolean | null;
  ApprovedFlag: boolean | null;
  AbsenceApprovedFlag: boolean | null;
  NoteReceivedFlag: boolean | null;
  AbsenceNoteReceivedFlag: boolean | null;
  ContactMadeFlag: boolean | null;
  AbsencePhoneCallMadeFlag: boolean | null;
  AbsenceLateArrivalFlag: boolean | null;
  AbsenceLateArrivalTime: string | null;
  AbsenceEarlyDepartureTime: string | null;
  StudentName: string;
  StudentNameInternal: string;
  StudentNameExternal: string;
  StudentLegalFullName: string;
  StudentSurname: string;
  StudentPreferred: string;
  StudentPreferredFormal: string;
  StudentFullFeeFlag: boolean;
  StudentYearLevel: string;
  StudentYearLevelSort: string;
  StudentCampus: string;
  StudentHouse: string;
  StudentForm: string;
  StudentTutor: string;
  StudentBoarder: string;
  StudentStatus: string;
  StudentBoarderType: string;
  StudentBoardingHouse: string;
  CurrentSemesterOnlyFlag: boolean;
  ClassCancelledFlag: boolean;
  NoteMadeFlag: boolean;
  AbsenceNoteMadeFlag: boolean;
  SubSchool: string;
}

export default iSynVAttendancesWithAbsence;